Fix renderPath method
There was a condition preventing the renderPath method being recursively
called with parameter x set to 0, resulting in missing moves coming from
the eighth row of the board (e.g. when the knight was on h7, g6 was
shown as four moves away instead of two, due to the move from f8 being
ignored). The most straightforward fix is removing the condition as it
does not seem necessary.